backend/requirements.txt
aiogram
aiogram_dialog
httpx
asyncpg~=0.30.0
Django~=4.2.19
djangorestframework
asgiref~=3.8.1
celery~=5.4.0
redis~=5.2.1
psycopg2-binary
aiohttp~=3.10.11
aiosignal~=1.3.2
frozenlist~=1.5.0
aiohappyeyeballs~=2.4.4
python-dotenv~=1.0.1
cachetools~=5.5.1
django-hashid-field
requests
environs